Add calculate endpoint

+++ src/main/java/ch/avelon/demo/persistence/MeasurementPoint.java
@@ -2,12 +2,9 @@ package ch.avelon.demo.persistence;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/**
- * @author Flurin Juvalta <flurin.juvalta@avelon.ch>
- */
 public interface MeasurementPoint {
     @JsonProperty("x")
-    long getRecordTime();
+    int getRecordTime();
     @JsonProperty("y")
     float getRecordedValue();
 }
